requirements.txt
gradio>=4.0
google-genai>=1.0
python-docx>=1.0
PyPDF2>=3.0
openai>=1.0
